Implemented player movement and first person camera
Added a player movement script to control player inputs from the keyboard.

Also added a player camera controller script so that the player can look around with their mouse.

Some additional changes include adding a ground layer, a script to connect the camera to the player, and a test scene with placeholder player and ground objects