Added monster movement, both based on player movement and camera movement
Made modifications to the PlayerMovement script to communicate with the monsterMovement script

Created a MonsterMovement script that is in charge of setting the monster's navmesh agent's position to the player's position, and moving the monster based on how much the player is currently moving their mouse or the player model.

I also set up a placeholder monster in the playerMovementTestScene for a visualization of what is happening.