Fixed Grapple Mechanic & Other Changes
Re imported all the prefabs for the game over screen and grapple UI, also hooked the grapple mechanic back up and set it to use Caleb's stun mechanic.

I also changed the modifier from that mechanic from 1.5 max speed to 1.15 max speed because it was really fast.

I also hooked the settings back button ui up and re added the grapplecomponent to the player and the gameover script to the gamemanager

I also added a check in the PauseMenu script to make sure the player couldn't enter the pause menu in the MainMenu or the GameOver screen