Added New model imports and made adjustments to grab code
Imported new models that can be used for the second level. In addition, changed some of the code in the grab functionality scripts. Some stuff still needs to be changed however. I would've liked to have done this is as a pr but we're running out of time and everyone else is asleep.